move AWS::S3::BucketPolicy
